Move InputWrapper to console.py
